implement basic sygus
